Improve invalid permission error handling
Instead of returning an error that stop execution when an
invalid permission identifier is requested, the permission
model will return a one line text indicating that the
permission name is invalid. This error message will be
displayed in place of the intended permission label.

This behavior was also extended to cover invalid permission
namespace requests.

A troubleshooting section is added explaining the possible
reasons for the and the solution.

Signed-off-by: Roberto Rosario <roberto.rosario@mayan-edms.com>