Document change type behavior change
Create a new permission to change the type of a document.

When support for changing the type of a document was added,
it was considered a property and controlled via the document
property edit permission.

Since changing the type of a documents now causes a cascade
of other changes, it was isolated as an individual class of
event along with its own permission.

The new document change type permission is required for the
document being changed and for the document type to which the
document will be changed into.

Signed-off-by: Roberto Rosario <roberto.rosario@mayan-edms.com>